# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/profile": {
      "patch": {
        "operationId": "ProfileController_patch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      }
    },
...
    "/api/v1/profile/photo": {
      "post": {
        "operationId": "ProfileController_uploadPhoto",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      },
      "get": {
        "operationId": "ProfileController_photo",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      }
    },
    "/api/v1/preferences": {
      "patch": {
        "operationId": "ProfileController_patchPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "Botvy",
    "description": "Commands are REST; reads are GraphQL; the live connection is a socket.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "service-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Token"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}